Finalized the multiple choice template for all modules
