+++ Exercises/binary_hex/files/mcquestion.py
@@ -1,8 +1,6 @@
 """
 key: The question number
-pair: This should be a list. indexes 0-3 are the choices. 4 is the answer. 5 is the question
-answer: This must be the same str value as the choice
-question: This is used in the header
+pair: [List of all the options, can be any length], Answer(This must be the same str value as the choice), Question
 """
 mc_dict = {
